new reward shaping result same with previous
